backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
praw==7.7.1
vaderSentiment==3.3.2
python-dotenv==1.0.1